+++ spec.go
@@ -3,8 +3,10 @@ package apptest
 import (
 	"context"
 
+	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/kubernetes"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 type Interface interface {
@@ -12,8 +14,14 @@ type Interface interface {
 	// and ensures they are installed by our app platform.
 	InstallApps(ctx context.Context, apps []App) error
 
+	// EnsureCRDs will register the passed CRDs in the k8s API used by the client.
+	EnsureCRDs(ctx context.Context, crds []*apiextensionsv1.CustomResourceDefinition) error
+
 	// K8sClient returns a Kubernetes clienset for use in automated tests.
 	K8sClient() kubernetes.Interface
+
+	// CtrlClient returns a controller-runtime client for use in automated tests.
+	CtrlClient() client.Client
 }
 
 type App struct {